Refactored the structure of the code
