+++ src/scenes/idle-game-scene/game-objects/market/resource-upgrade-market-groups/ResourceUpgradeMarketGroupComponent.ts
@@ -1,8 +1,8 @@
 import '../../../../../ui/LuuButton';
 import LuuButton from '../../../../../ui/LuuButton';
 import Upgrade from "../../../upgrades/Upgrade";
-import ResourceUpgradeManager from "../../../upgrades/upgrade-managers/ResourceUpgradeManager";
-import CollectSpeedUpgrade from "../../../upgrades/upgrade-types/CollectSpeedUpgrade";
+import ResourceUpgradeManager from "../../../upgrade-managers/resource-upgrade-managers/ResourceUpgradeManager";
+import CollectSpeedUpgrade from "../../../upgrades/resource-upgrades/CollectSpeedUpgrade";
 import UpgradeGameObject from "../../UpgradeGameObject";
 
 export class ResourceUpgradeMarketGroupComponent extends UpgradeGameObject {
