+++ src/scenes/idle-game-scene/upgrade-managers/resource-upgrade-managers/CoalUpgradeManager.ts
@@ -1,11 +1,11 @@
-import Log from "../../resources/Log";
+import Coal from "../../resources/Coal";
 import ResourceManager from "../../resources/resource-managers/ResourceManager";
-import Upgrade from "../Upgrade";
-import CollectSpeedUpgrade from "../upgrade-types/CollectSpeedUpgrade";
+import Upgrade from "../../upgrades/Upgrade";
+import CollectSpeedUpgrade from "../../upgrades/resource-upgrades/CollectSpeedUpgrade";
 import ResourceUpgradeManager from "./ResourceUpgradeManager";
 
-export default class LogUpgradeManager extends ResourceUpgradeManager<Log> {
-	public constructor(resourceManager: ResourceManager<Log>) {
+export default class CoalUpgradeManager extends ResourceUpgradeManager<Coal> {
+	public constructor(resourceManager: ResourceManager<Coal>) {
 		super(resourceManager);
 
 		this.initializeUpgrades();
@@ -17,7 +17,7 @@ export default class LogUpgradeManager extends ResourceUpgradeManager<Log> {
 		for(let i = 0; i < 10; i++) {
 			let upgradeCost = Math.round(baseCost * Math.pow(2.65, i));
 			let upgradeSpeed = Math.round(baseSpeed * 1.8);
-			this.upgrades[Upgrade.Type.COLLECT_SPEED].enqueue(new CollectSpeedUpgrade(i + 1, 'LOG_COLLECT_SPEED_INCREASE_' + i, upgradeSpeed, upgradeCost));
+			this.upgrades[Upgrade.Type.COLLECT_SPEED].enqueue(new CollectSpeedUpgrade(i + 1, 'COAL_COLLECT_SPEED_INCREASE_' + i, upgradeSpeed, upgradeCost));
 		}
 	}
