--- src/scenes/idle-game-scene/upgrades/upgrade-types/UpgradeType.ts
@@ -1,3 +1,0 @@
-export default class UpgradeType {
-	
-}
